DjangoAssignment/DjangoTwitterAssignment/requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11
django-filter==1.1.0
django-filters==0.2.1
djangorestframework==3.7.3
idna==2.6
oauthlib==2.0.6
pkg-resources==0.0.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.5.0
urllib3==1.22
